Try to match the whole part name
Problem: We are greedy when matching part names. If we have `foo` and
`foobar` available, and we want to parse `foobar`, we'll match the first
name part that we see. I.e., we may match it as `foo` if we see first!

Solution: Do not only match the suffix, but also check whether there
are no other characters that follow it (i.e., we have spaces). The side
effect is that things like parentheses, which are implemented as name
parts, now require spaces around them, i.e., `( foo )` instead of
`(foo)`.